+++ src/gamestate_savegame.cpp
@@ -350,7 +350,7 @@ SetupHeaderInfo readSetupHeader(const nlohmann::ordered_json &j)
 	game.playerLeaveMode = static_cast<PLAYER_LEAVE_MODE>(jopt.at("playerLeaveMode").get<uint8_t>());
 	game.playerReconnectWaitSeconds = clampPlayerReconnectWaitSeconds(jopt.at("playerReconnectWaitSeconds").get<uint32_t>());
 	// default to the legacy A* planner for snapshots written before this setting existed
-	game.pathfindingBackend = jopt.value("pathfindingBackend", static_cast<uint8_t>(0));
+	game.pathfindingBackend = jopt.value("pathfindingBackend", static_cast<uint16_t>(0));
 
 	readPlayers(j.at("players"));
 	readStructureLimits(j.at("structureLimits"));
